03_ner_prj/working/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
beautifulsoup4==4.6.0
boto3==1.17.59
botocore==1.20.59
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gluonnlp==0.10.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.32.0
h5py==2.10.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
JPype1==1.2.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.9.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==9.0.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-crf @ git+https://github.com/kmkurn/pytorch-crf@e843718540512eb04d9b756fca04a0915affe175
pytorch-transformers==1.2.0
pytz==2021.1
regex==2021.4.4
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.3
sentencepiece==0.1.95
six==1.15.0
tb-nightly==2.6.0a20210427
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.2
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.5.1
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1